任务启动和重启加锁，加入状态判断
